fast-api-react/backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.0"
description = "Easily run background tasks in FastAPI with Hatchet"
authors = []
readme = "../README.md"

[tool.poetry.scripts]
api = "src.api.main:start"
hatchet = "src.workflows.main:start"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
fastapi = "^0.109.0"
openai = "^1.11.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
urllib3 = "2.2.1"
hatchet-sdk = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
